requirements.txt
pytesseract
streamlit
torch
pillow
transformers
opencv-python
numpy
